Simplify apt jar and dependencies. (#357)
See https://github.com/google/auto/tree/master/common#processor-resilience

The basic idea is to keep the apt jar as simple as possible, and
avoid external depenencies or classes that might conflict on the
classpath of the code being compiles with this processor. On Gradle
or some other build tool we can rely on classpaths not getting mixed,
but it is riskier in maevn, so we need to take care not to have
some library version in this processor break another processor or
vice versa, or to have some dependency here force another library
to be updated or vice versa.