Add (partial) generics specialization
Domino UI lacks generics specialization in many places,
but this commit focuses on two classes:

* usage of elemento.IsElement
* BaseDominoElement implementation

Modification on some unrelated classes was needed too,
to make method signatures distinct enough for compiler.

Note that this could breake some projects which used @Override on
old (unspecialized) method variants.

This is NOT complete fix for lack of use of generics in Domino UI.
But it's a start.